Fixed negative temperatures

+++ devicetypes/smartthings/smartsense-temp-humidity-sensor.groovy
@@ -180,7 +180,7 @@ private Map parseCustomMessage(String description) {
 }
  
 def getTemperature(value) {
-	def celsius = Integer.parseInt(value, 16) / 100
+	def celsius = Integer.parseInt(value, 16).shortValue() / 100
 	if(getTemperatureScale() == "C"){
 		return celsius
 	} else {
